Moved tests and acceptance tests to test directory.
