+++ test/bugsplat.acceptance-spec.ts
@@ -2,7 +2,7 @@ import { TestBed, async } from '@angular/core/testing';
 import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
 import { HttpClientModule, HttpClient } from "@angular/common/http";
 import { RequestOptions } from "@angular/http";
-import { BugSplat } from "./bugsplat";
+import { BugSplat } from "../src/bugsplat";
 
 const testUser = "Fred";
 const testPassword = "Flintstone";
